net: golioth: do not hold lock when connecting
Connect is a relatively long operation, so keeping mutex for that long
does not seem to be a good idea. This for example resuts in blocking all
golioth consumers that try to send a packet.

Just start connecting to golioth without acquiring lock. Then, if
connect was successful, just write socket number while holding lock for
very short time.