samples: hello: trigger reconnect on receive timeouts
So far there was only single point during initialization when
application tried to connect to golioth. This was good enough for stable
connections, such as QEMU+PPP, but it is not enough for real life use
cases. Add 30s RX timeout, which will trigger reconnect whenever there
is no incoming traffic.