README: remove `west patch` step
This command is not required for a long time already. Remove it from
README in order not to confuse users. Leave it in CI (GitHub and
GitLab), as it is handy to be able to test new features from Zephyr from
time to time (without intension of merging those patches to main
branch).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>