logging: golioth: fix error handling in message processing
In case of logging v1 there was a possibility that log_msg_put()
was *not* called if an error occurred during processing. This resulted
in log message leak.

In case of logging v1 and v2 message logging index was not incremented
in case of error, so make sure to do it correctly now.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>