README: update links to Zephyr RTOS documentation to point to 3.0.0
So far links to 'latest' Zephyr RTOS documentation were used. All those
are broken at this moment except link to 'ESP32' board.

Update all of those (including ESP32 board link, for consistency) to
version 3.0.0 of Zephyr RTOS. Currently the recommended Zephyr RTOS
version (the one pointed to from `west.yml`) is somewhere between 2.7.0
and 3.0.0, so the latter is the best candidate as of now.

In future all links to Zephyr RTOS documentation (and any other
versioned documentation) need to be adjusted together with the component
version that is used.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>