settings: enable flash in nrf9160dk_nrf9160_ns.conf
When testing mcumgr in the settings sample on the nrf9160dk_nf9160_ns
board, the following error is received:

$ mcumgr \
    --conntype=serial \
    --connstring='dev=/dev/ttyACM0,baud=115200' \
    config golioth/psk-id nicks_esp32s3_devkit-id@nicks-first-project
Error: 1

This is happening because the board config does not have CONFIG_FLASH
enabled. This is a required setting in order for MCU manager to be able
to save settings to flash.

The other board files, like nrf52840dk_nrf52840.conf, already have flash
enabled, so it seems to just be an oversight for this particular board.

Signed-off-by: Nick Miller <nick@golioth.io>