Add file west-ncs.yml
This is intended to be used in scenarios where our SDK is the manifest
repo, as described in NCS's Workflow 4 ("app as the manifest repo").

Previously, we were manually editing NCS's
west.yml to include the Golioth SDK, but this was manual and
error-prone, and it was too easy to select an invalid/non-working
version of NCS.

Adding west-ncs.yml allows us to use NCS the same way we use vanilla
Zephyr for cases where the SDK is used as the manifest repo. The
version of NCS is pinned in this .yml, which avoids external users
having to manually select the NCS version, or add the Golioth SDK
as a module in NCS's west.yml.

Signed-off-by: Nick Miller <nick@golioth.io>