net: golioth: fix golioth_sendmsg() return code check
Return code was saved to 'ret' variable, but 'err' variable was checked in
the next like instead.

In fact number of sent bytes should be compared with the number returned
from underlying zsock_send() call. Do exactly that inside
golioth_sendmsg(), as that function contains number of total bytes that
should be sent. Change golioth_sendmsg() behavior to return 0 in case of
success and negative value in case of error.

Save value returned from golioth_sendmsg() to 'err', which by convention
can be 0 (success) or negative (error).

Reported-by: Nick Miller <nick@golioth.io>
Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>