net: golioth: enforce DTLS handshake in golioth_connect()
As per `connect(3)` manpages:

  The connect() function shall attempt to make a connection on a
  connection-mode socket or to set or reset the peer address of a
  connectionless-mode socket.

and:

  If the initiating socket is connection-mode, then connect() shall attempt
  to establish a connection to the address specified by the address
  argument. If the connection cannot be established immediately and
  O_NONBLOCK is not set for the file descriptor for the socket, connect()
  shall block for up to an unspecified timeout interval until the
  connection is established.

RFC 6347 (DTLS 1.2) interchangably uses the terms association, session and
connection. Though it is not specified in neither POSIX nor Zephyr how DTLS
sockets should be treated (they are really just Zephyr extension), it seems
natural to assume they are "connection-mode" sockets as per `connect()`
POSIX documentation.

Because of the reasons described above, make sure that golioth_connect()
does the DTLS handshake in order to "establish a connection". As Zephyr's
connect() call does not result in such behavior, issue send() with an empty
CoAP packet, so that the DTLS handshake is done as part of that API call.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>