west-zephyr: bump MCUboot to include fix for ESP32
ESP32 platform had broken chain-loading of application image at least
starting from Zephyr 2.7.0 stable release and it was still not fixed with
the release of 3.1.0.

In fact the problem is in MCUboot bootloader, but it is Zephyr which
decides what version of MCUboot is pulled in.

Change revision of MCUboot that is used with west-zephyr.yml manifest file
to a version that contains a fix for ESP32 application chain-loading.
Fortunately used revision contains only this single fix (compared to
revision pulled in by Zephyr 3.1.0), so there is no risk of breaking other
functionalities/platforms.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>