net/golioth: add more on_message callbacks
For observations, in order for the observation packets to
be received, the coap_resonse_received function must be
called. Typically this is done inside of the golioth_client
on_message callback, which is set by users.

For the upcoming golioth_rpc feature, the
plan is to observe on path ".rpc" inside the SDK, not in the
application. There's only one on_message callback (which
is already reserved for the user), so it makes it difficult
to observe inside the SDK, since there's no straightforward way
to call coap_response_received.

This commit adds storage for additional on_message callbacks
inside golioth_client, leaving the old on_message callback
intact for backward compatibility.

For the future, these additional on_message callbacks could be
used to improve the public APIs so that users aren't forced
to set on_message and call coap_response_received (this could
be hidden inside the SDK).

Signed-off-by: Nick Miller <nick@golioth.io>