Add golioth_rpc feature
Design choices:

* Use CBOR for the payload. Originally had planned to
use the Zephyr JSON library, but ran into issue with
parsing NULL params and dealing with the params array in
general, since we don't know ahead of time what the types
of the items in the params array will be. Considered depending
on cJSON, as it's more flexible, but it turns out that QCBOR
is equally as flexible and already a dependency, so decided
to go with QCBOR.

* Hide the feature behind a Kconfig feature flag.
Disabled by default since the feature is not yet deployed
on the cloud side.

* Hide the observation and coap_reply inside golioth_rpc.c
so the user doesn't have to manage it.

Signed-off-by: Nick Miller <nick@golioth.io>