samples: shell: new sample
Introduce a sample that exposes low-level SDK APIs as shell commands. This
works below (i.e. does not use) "golioth_system_client" module, so that
there is full control over DTLS configuration (configuring PSK-ID, PSK or
certificates) and connection management (no logic, just
connect/disconnect/is_connected commands).

This is supposed to be used for testing and verifying low-level APIs as
well as low-level mechanisms, such as DTLS handshakes and certificate
verification. The next step will be low-level CoAP mechanisms verification,
like retransmission of confirmed packets and testing application callbacks.

Passing DTLS PSK-ID and PSK is straightforward, as they are usually short
enough to be passed as shell command argument. However passing certificates
is more cumbersome, as it is usually few hundreds of bytes and shell over
UART has a limit of characters that can be passed at once. This means that
in most cases (most UART devices) it is needed to send data in chunks
instead of all at once.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>