Add golioth_settings feature
This is a thin layer that interfaces with the Golioth Settings
service, handling the CoAP observation of ".c" and sending
the response to ".c/status".

User code has to register a callback which will process a
single setting, and the library will call the callback for
each setting received from Golioth.

Added example code to the existing "settings" sample instead
of creating a new one, to avoid growing the already-extensive
list of samples.

Signed-off-by: Nick Miller <nick@golioth.io>