{rpc,settings}: forward declare qcbor typedefs and drop header use
Currently both RPC and Settings headers pull QCBOR headers. This means that
QCBOR module needs to be added as Zephyr module (e.g. by specifying in west
manifest when west is used for repository management). This is true even
when neither RPC nor Settings are actually used (like it is the case for
simplest 'hello' sample).

Remove QCBOR header inclusion and forward declare QCBORDecodeContext and
QCBOREncodeContext types instead. As pointers to those types are used in
the Golioth RPC/Settings APIs, then there is no need to have full
information about those types as long as their values are not referenced.

This solves build errors of samples that do not use RPC and Settings and do
not enable CONFIG_QCBOR.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>