samples: add test sample
Adding a new sample named test which is intended to run
automated tests in CI.

Currently, it only supports nrf52840dk, and has a single test
to verify that the board can connect to Golioth.

Another workflow was added for the nrf9160dk, but is currently
disabled due to inconsistency in modem initialization which
can cause the test to fail. It will be re-enabled once the
initialization issue is fixed (JIRA: DSDK-334).

Signed-off-by: Nick Miller <nick@golioth.io>