settings: option to persist Golioth settings
Settings will be stored to persistent storage (via Zephyr's
Settings subsystem), but only if the user's settings
callback returns success, indicating the setting is
known and valid.

The intended use is for users that
need access to their settings before connecting to
Golioth. It's up to the user to retrieve the settings
out of the Zephyr Settings subsystem, if desired.

Signed-off-by: Nick Miller <nick@golioth.io>