net/golioth: remove hexdump of tx packet
If CONFIG_GOLIOTH_LOG_LEVEL_DBG is set to 'y',
then the system will freeze at boot after a couple
of seconds (tested on nRF9160dk).

The reason why this log message will result in freeze is
that for each generated log message, we will once
again enter the same code path when trying to send
the log message to Golioth. This results in an infinite
log loop, which creates a kind of deadlock.

There may be other occurrences of this kind of deadlock
in the code, but testing has shown that removing this
log statement results in more stable logging (no deadlock).

Signed-off-by: Nick Miller <nick@golioth.io>