golioth.c: workaround note for using IP in X509
When testing with a local golioth instance and a certificate
with Subject Alternative Name set to a hard-coded IP address
(not a DNS name), hostname verification will fail.

Added a note about a workaround to disable hostname
verification in this case.

This will usually only be needed for Golioth devs testing
locally, so it doesn't make sense to put in Kconfig as an
option.

Signed-off-by: Nick Miller <nick@golioth.io>