samples: lightdb*: use 'json' instead of 'text/plain'
Use of 'text/plain' content type is not technically correct when
settings/observing number types ('counter' which is an integer, 'temp'
which is a floating-point number) in lightdb. 'text/plain' could
potentially be used with text values only. The main reason why this is
interpreted as number by server side is by falling back to 'json' from
unsupported 'text/plain' content type in server implementation.

Use 'json' content type, so that it makes clear that a string
representation of number is used (values are passed as 'NUM', whereas a
string would be passed as '"NUM"').

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>