golioth: introduce PATHV() helper macro
This macro allows to create valid 'const uint8_t **pathv' with sentinel at
the end. The result of PATHV() macro is Compound literal [1], which has
very useful property (according to [1]):

  The value category of a compound literal is lvalue (its address can be
  taken).

So following:

  const uint8_t *pathv[] = {
          GOLIOTH_FW_REPORT_STATE,
          package_name,
          NULL, /* sentinel */
  };

  golioth_some_api(..., pathv, ...);

can now be replaced with:

  golioth_some_api(..., PATHV(GOLIOTH_FW_REPORT_STATE, package_name), ...);

Main advantage of using this macro instead of defining array explicitly is
that sentinel is appended at the end automatically. This prevents certain
user errors, which could result in memory access errors (with missing NULL
termination). Another advantage, even more obvious one, is more dense
code (hopefully with the same level of readability).

[1] https://en.cppreference.com/w/c/language/compound_literal

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>