samples: lightdb_stream: update to new API
Demonstrate use of new APIs:

 * golioth_stream_push(),
 * golioth_stream_push_cb().

Implement helper functions temperature_push_*() for both, so that it is
clear in main() what type of communication (synchronous or asynchronous)
happens there.

Update documentation with new logs that are printed by this sample.
Additionally add output of 'goliothctl stream listen', which shows data in
realtime.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>