samples: lightdb_led: use ON/OFF instead of 1/0 in log message
It is quite clear when there are 4 LEDs configured (as documented in
README.rst) what each number in "LED %d -> %d" means. However when there is
only one LED, then "LED 0 -> 0" message is not that clear.

Use "ON" and "OFF" instead of "1" and "0", so it makes it clear to the user
that second value is actually desired (configured) state of LED.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>