.github: assume JLink SNR set in environment
If the self-hosted runner has multiple JLink devices, then
we can no longer rely on `$(nrfjprog --ids)` to obtain
the correct serial number of the board.

Instead, the self-hosted runner admin will have to set
the serial number in the environment, otherwise programming
of nRF52 and nRF91 boards may fail.

Signed-off-by: Nick Miller <nick@golioth.io>