samples: common: settings_shell: use shell_{print,warning,error}
So far there was no consistence around using:
 * shell_error(...) vs
 * shell_fprintf(..., SHELL_ERROR, ...) vs
 * shell_fprintf(..., SHELL_VT100_COLOR_RED, ...)

as well as with:

 * shell_print(...) vs
 * shell_fprintf(..., SHELL_NORMAL, ...) vs
 * shell_fprintf(..., SHELL_VT100_COLOR_GREEN, ...)

Unify those and always use shell_error() in the first case and
shell_print() in the latter case. Note that this changes color of some
printed strings from green to default (white/grey), but this was done on
purpose (to improve consistency of printed strings across other Zephyr
shell commands, that also use default font color in most cases).

Change also single invocation of shell_fprintf(..., SHELL_WARNING, ...) to
shell_warn().

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>