samples: common: settings_shell: add shell_json_{print,error}
Introduce helpers that allow to deduplicate some boilerplate related to
printing in JSON or in plain text.

Format string are now the same for JSON and plain text, so resulting code
is less prone to copy-paste errors.

Note that all format/message strings now start with uppercase for
consistency across JSON and plain text output.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>