samples: dfu: disable settings in system client for NCS runtime test
Disable settings support in system client, so that PSK-ID and PSK will be
pulled in from environment variables and baked into the firmware. This
allows to omit credentials provisioning.

This overlay is only for NCS runtime tests in twister for now. This
should be applied to sysbuild-based DFU build, but unfortunately it
seems not supported yet (by adding those options to sample.yaml).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>