samples: common: wifi: handle DISCONNECT_RESULT wifi_mgmt event
So far only CONNECT_RESULT wifi_mgmt event was handled. This should be
enough with properly functioning WiFi drivers, which should use that event
for either succeeded or failed connect attempt (status code is always
passed with CONNECT_RESULT, which defines whether it was a success or
failure). However current implementation of ESP32 WiFi driver seems to not
respect wifi_mgmt APIs and sends DISCONNECT_RESULT wifi_mgmt result on
failed connect attempt.

Handle DISCONNECT_RESULT events and treat that as if CONNECT_RESULT was
received with -EFAULT status code was received. This workarounds current
ESP32 WiFi driver implementation. This driver needs to be fixed anyway, but
more drivers might appear in the future with such broken behavior.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>