samples: lightdb/get: pytest: fix script to expect hexdump of counter
Counter values are logged now using hexdump, so replace regex matching of a
number with fetching counter value from a hexdump.

This was not detected in GitLab CI due to limited integration between
twister and pytest. The only noticeable difference was in twister logs in
CI, which were like:

  samples/lightdb/get/sample.golioth.lightdb_get.fast PASSED (qemu)

instead of:

  samples/lightdb/observe/sample.golioth.lightdb_observe.fast PASSED \
    (qemu 17.617s)

(there is no execution time printed in case of failed lightdb/get tests).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>