test: connect to coap.golioth.dev, instead of .io
The intent is to run the end-to-end hardware test against
the Golioth dev server, to catch incompatibilities/issues
before the changes hit prod.

Added comments to the CI workflow files to make it clear
that the device credentials should come from the dev server.

Outside of the repo, the credentials YAML file was updated
on self-hosted runners.

Signed-off-by: Nick Miller <nick@golioth.io>