.github: run hardware tests on a nightly schedule
Previously, the hardware tests would only run in CI when
a PR was created or merged to main.

There are changes/features/fixes pushed to Golioth dev
servers once or twice a week, each of which have the potential
to cause an incompatibility between client and server.

We want to catch these kinds of issues sooner
rather than later, even if there aren't any active PRs.
By running the hardware tests on a nightly schedule,
we ensure a worst-case detection time of 24 hours.

Signed-off-by: Nick Miller <nick@golioth.io>