golioth: drop MBEDTLS_CIPHER_GCM_ENABLED selection
This option was selected during early development. As of today it is not
needed, as CBC ciphersuites are enabled by default in mbedTLS and
TLS_PSK_WITH_AES_128_CBC_SHA256 is supported by Golioth server. There is no
point in enforcing GCM algorithm to be built into the firmware when there
is working alternative.

Enable MBEDTLS_CIPHER_GCM_ENABLED only when the other two (CCM and CBC) are
not enabled already. Do this with `imply`, so that there is no circular
dependency and there is always a possibility to disable it when not
desirable.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>