scripts: python: golioth: raise exceptions on WebSocket errors
When errors are received over WebSocket, they are are packed into JSON
objects with 'error' member. Detect such case and raise an exception if
that happens, so that caller can easily handle it.

It is not clear what is the meaning of codes reported in
{"error":{"code":code}} JSON message, but it is likely that those match
RPCStatusCode enums. Use those for now and revisit at some point in future
when that clarifies.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>