tests: lightdb: propagate callback error to async.err
If there was an error during asynchronous (callback-based) request, then
such error was not propagated to `struct async_data` structure, but instead
it was still set to `0`, just like the API would be successful. Later, test
would discover a problem anyway, as received data would not match with
expected one. However, it is better to propagate error and not rely on data
comparison.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>