settings: add new API golioth_settings_observe()
Before this change, when `golioth_settings_register_callback()` is called,
an observation is automatically established.

However, this “automatic observe” behavior has several issues:

1. It might be unwanted in some user applications (e.g. low-power/low-data)
that don’t want the overhead of observations

2. It’s inconsistent with the way observations work for DFU.
For DFU, it’s expected that the user calls golioth_fw_observe_desired()
in the on_connect callback (i.e. observations are not automatically
established).

3. It can be unexpected/surprising that an API named “register”
would also create an observation.

This commit separates settings registration from observation, and makes the
observation an explicit API call.

This is a breaking change to user applications, as they are now required
to call golioth_settings_observe() in the on_connect callback if they wish
to observe settings. The settings sample was updated with this change.

Signed-off-by: Nick Miller <nick@golioth.io>