rpc: add new API golioth_rpc_observe()
Before this change, when `golioth_rpc_register()` is called, an
observation is automatically established if it’s the first RPC method
being registered.

However, this “automatic observe” behavior has several issues:

1. It might be unwanted in some user applications (e.g. low-power/low-data)
that don’t want the overhead of observations

2. It’s inconsistent with the way observations work for DFU.
For DFU, it’s expected that the user calls golioth_fw_observe_desired()
in the on_connect callback (i.e. observations are not automatically
established).

3. It can be unexpected/surprising that an API named “register”
would also create an observation.

As a separate issue, the prior behavior of only observing on the first
registered RPC method was flawed, since if a reconnect ever happens
then the observation is never established again. This commit removes
that behavior.

This commit separates RPC registration from observation, and makes the
observation an explicit API call.

This is a breaking change to user applications, as they are now required
to call golioth_rpc_observe() in the on_connect callback if they wish
to observe RPCs. The rpc sample was updated with this change.

Signed-off-by: Nick Miller <nick@golioth.io>