.github: add DFU testing workflow using Twister
This commit adds a workflow to build and test the DFU sample on the
nRF52840dk using Twister on a self-hosted runner.

Different WiFi credentials, device credentials, device serial numbers, and
device ports are needed to run a Twister workflow on one of several
self-hosted runners. This workflow uses the runner name to select the
correct GitHub secrets for the devices.

Use this formatting for converting the runner name to a prefix for the
GitHub Secret key:

* Convert dashes to underscores
* Convert lowercase to uppercase

Three GitHub Secrets are needed to provide credentials and device
specifics:

* <SELF_HOSTED_RUNNER_NAME>_NRF52840DK_PORT
* <SELF_HOSTED_RUNNER_NAME>_NRF52840DK
* GOLIOTHCTL_YAML

Contents and formatting of these file are detailed in the comments of the
workflow file. Use base64 to encode the files for storage as GitHub Secrets

$ base64 -i --wrap=0 file-to-encode.yml

Runners indicate their ability to perform this workflow by adding the
`has_twister_nrf52840` tag.

Notes:

* Tests that use Twister should be prefaced with `twister_` to indicate
  they need a larger self-hosted runner than more basic HIL testing.
* Golioth self-hosted runner install instructions have been updated to
  install SEGGER tools to /opt/SEGGER/JLink/ directly. This fixes a
  previous challenge where symlinks were used on the runner, but could not
  be mounted into a Docker container when running Twister.

Signed-off-by: Mike Szczys <mike@golioth.io>