samples: dfu: remove CONFIG_BOOTLOADER_MCUBOOT from prj.conf
This configuration option was set, so that `west build ...` invocation
would already build bootloader-compatible (meaning there is resevered space
at the beginning of the flash for bootloader use) binaries.

Since sysbuild is applying that option automatically when invoking `west
build --sysbuild ...`, there is no need to have it listed in `prj.conf`.

Enable CONFIG_BOOTLOADER_MCUBOOT=y in platforms requiring NCS (i.e. nRF9160
DK), as that is where sysbuild configuration is not applied (because it is
not used).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>