samples: dfu: pytest: remove global NEW_VERSION
Do not hardcode new firmware version (2.0.0) globally and move it into the
test itself. This will allow to develop more DFU tests, that allow to test
a flow with multiple rollouts (thus multiple versions).

Introduce a Firmware helper class, which will store both the path and the
version of the firmware. Use 'imgtool' package in order to figure out the
version automatically from file name (in the context where only a firmware
object is passed).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>