.github: workaround when missing initial IDLE log
During testing, add a 10-second delay when first booting to give the
self-hosted runner enough time to set up log monitoring.

The Twister pytest uses WebSockets to monitor the firmware state messages
sent to Golioth. When run as a workflow on a self-hosted runner, the time
between flashing the binary and monitoring the logs is enough that the
initial IDLE state is often missed, resulting in a false negative.

Signed-off-by: Mike Szczys <mike@golioth.io>