ci: replace 'git apply' with 'git am' when applying patches
So far `west patch --apply` was used in order to apply patches without
creating commits for each specific patch. This was a workaround due to
unconfigured git user.{email,name} options in CI environment.

Use `west patch` which calls `git am` (instead of `git apply`) underneath,
so that commit messages are created for each applied patch. This gets rid
of situation when there are unstages changes in repository and then this
state is cached by CI for the next CI run. As a result, when caching was
enabled, it was required to clean unstanged changes with `git clean ...`
and `git reset ...` invocations for each west project. By switching to `git
am` this is no longer an issue.

Configure git user.{email,user} options to some dummy values. Those will
not be visible publicly, but will be used only internally during the CI
run. This makes sure that `git am` will run without issues.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>