bump Zephyr to 3.4.0+
MINIMAL_LIBC_MALLOC_ARENA_SIZE got deprecated in favor of
COMMON_LIBC_MALLOC_ARENA_SIZE, so update that.

Rework pytest testcases to use `dut` fixture, which is needed in order to
flash device before running test.

Tip of main (as opposed to 3.4.0 release) contains eventfd related fixes,
which are needed in order to wakeup poll() using eventfd_write().

--cmdopt is no longer passed to pytest invocation from twister. Use
request.config.build_dir instead.

Add timeouts to sample.yaml configurations, as from now on pytest twister
harness respects that timeout (which was not entirely true with the initial
pytest support).

Use 'hal_espressif' revision that contains fix for duplicated enum. This is
just a temporary solution, as this revision is not used yet in Zephyr's
west manifest.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>