kconfig: add options for maximum PSK-ID and PSK lengths
PSK-ID length was always hardcoded to maximum 64 characters. As a result
long project names with long PSK-IDs (due to
'long-device-id@long-project-name' PSK-ID format) would not be supported.

PSK had the same issue, but only when used with NCS.

Introduce options for both, PSK-ID and PSK, so that long credentials can be
supported without changing SDK codebase.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>