auth: remove hardcoded certificates from golioth library
Clients are now responsible for setting certificates in the TLS Credential
subsystem before starting Golioth

samples: add support for hardcoded certs in hello sample

The hello sample now shows a trivial example of loading certificates from
files selected at build time. This does not demonstrate provisioning, but
is the simplest way to show clients how to load certificates into the
appropriate storage module so that the TLS library can find and use them.
It also maintains the same ease-of-use for samples and experimentation.

Signed-off-by: Sam Friedman <sam@golioth.io>