workflows: add recovery for nRF devices
Nordic devices can sometimes enter an access protection mode, causing HIL
testing to encounter an "Access protection is enabled, can't read device
version." error that prevents flashing and causes CI to fail. This commit
adds the `--recover` operation to ensure the board is in a flashable state
for each test.

Signed-off-by: Mike Szczys <mike@golioth.io>