bump Zephyr to v3.5.0-rc1 and NCS to tip of main
v3.5.0-rc1 is the newest release candidate. Update now, so there are more
chances to discover potential bugs.

Bump NCS to tip of main, so that twister implementation is almost on par
with vanilla Zephyr 3.5.0-rc1. Drop old cherry-picked patches and apply 3
newer. This allows to keep the same pytest test scripts for both vanilla
Zephyr and NCS.

Drop LTE_LC_NW_REG_REGISTERED_EMERGENCY in switch-case, as it was dropped
in NCS as well.

Bump Zephyr SDK for NCS, as it requires newer version of 'west'.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>