net: enable extended master secret
Extended Master Secret is a DTLS extension that ensures the
Master Secret is tied to the handshake parameters and is used to
prevent certain Man in the Middle attacks. Golioth reserves the right
to enforce the use of Extended Master Secret on the server side in
the future, so this option must be enabled to ensure continued
compatibility with Golioth.

Signed-off-by: Sam Friedman <sam@golioth.io>