samples: dfu: add log for different release number
Add message when a release number is received from Golioth that doesn't
match what is currently running. This indicates an OTA firmware update
should be performed (it will either be an update or a rollback). This
change compliments the message we log when the received semantic version
matches what is currently running on the device.

Signed-off-by: Mike Szczys <mike@golioth.io>